New structure started, separate tests from src.
