+++ DotNetCross.Memory.Unsafe.sln
@@ -1,25 +1,24 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DotNetCross.Memory.Unsafe", "DotNetCross.Memory.Unsafe\DotNetCross.Memory.Unsafe.xproj", "{A421B9BA-2399-4B78-A4A8-02AD3A6CF2F2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetCross.Memory.Benchmark", "DotNetCross.Memory.Unsafe.Benchmark\DotNetCross.Memory.Benchmark.csproj", "{335FFEC5-1982-4B23-A3DB-71EACCFE18E6}"
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DotNetCross.Memory.Unsafe.UnitTests", "DotNetCross.Memory.Unsafe.UnitTests\DotNetCross.Memory.Unsafe.UnitTests.xproj", "{2E86AFF9-7EBC-491E-A178-2C73403836ED}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DotCross.Memory.Unsafe", "DotCross.Memory.Unsafe", "{813110EC-26A8-4273-A22F-EBE8507834AF}"
 	ProjectSection(SolutionItems) = preProject
 		..\AssembleRewrittenIL.bat = ..\AssembleRewrittenIL.bat
 		..\BuildBase.bat = ..\BuildBase.bat
 		..\DisassembleBase.bat = ..\DisassembleBase.bat
 		..\il\DotNetCross.Memory.Unsafe.BASE.il = ..\il\DotNetCross.Memory.Unsafe.BASE.il
-		..\build\DotNetCross.Memory.Unsafe.nuspec = ..\build\DotNetCross.Memory.Unsafe.nuspec
 		..\il\DotNetCross.Memory.Unsafe.il = ..\il\DotNetCross.Memory.Unsafe.il
+		build\DotNetCross.Memory.Unsafe.nuspec = build\DotNetCross.Memory.Unsafe.nuspec
+		global.json = global.json
 		..\NugetPack.bat = ..\NugetPack.bat
 		..\README.md = ..\README.md
 	EndProjectSection
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DotNetCross.Memory.Unsafe", "src\DotNetCross.Memory.Unsafe\DotNetCross.Memory.Unsafe.xproj", "{A421B9BA-2399-4B78-A4A8-02AD3A6CF2F2}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DotNetCross.Memory.Unsafe.UnitTests", "tests\DotNetCross.Memory.Unsafe.UnitTests\DotNetCross.Memory.Unsafe.UnitTests.xproj", "{2E86AFF9-7EBC-491E-A178-2C73403836ED}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,10 +30,6 @@ Global
 		{A421B9BA-2399-4B78-A4A8-02AD3A6CF2F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A421B9BA-2399-4B78-A4A8-02AD3A6CF2F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A421B9BA-2399-4B78-A4A8-02AD3A6CF2F2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{335FFEC5-1982-4B23-A3DB-71EACCFE18E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{335FFEC5-1982-4B23-A3DB-71EACCFE18E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{335FFEC5-1982-4B23-A3DB-71EACCFE18E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{335FFEC5-1982-4B23-A3DB-71EACCFE18E6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2E86AFF9-7EBC-491E-A178-2C73403836ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2E86AFF9-7EBC-491E-A178-2C73403836ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2E86AFF9-7EBC-491E-A178-2C73403836ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
