--- src/DotNetCross.Memory.Unsafe.UnitTests/project.json
@@ -1,31 +1,0 @@
-{
-    "version": "0.2.3",
-
-    "compilationOptions": {
-        "allowUnsafe": true
-    },
-
-    "dependencies": {
-        "DotNetCross.Memory.Unsafe": "0.2.3.2",
-        "Microsoft.NETCore.Runtime": "1.0.2-rc3-24112-00",
-        "xunit": "2.2.0-beta1-build3239",
-        "xunit.runner.console": "2.2.0-beta1-build3239"
-    },
-
-    "frameworks": {
-        "dnxcore50": {
-            "dependencies": {
-                "xunit.runner.dnx": "2.1.0-rc1-build204"
-            }
-        },
-        "netcore50": {
-            "dependencies": {
-                "System.Runtime": "4.1.0-rc3-24112-00"
-            }
-        }
-    },
-
-    "commands": {
-        "test": "xunit.runner.dnx"
-    }
-}
