+++ src/DotNetCross.Memory.Unsafe/DotNetCross.Memory.Unsafe.xproj
@@ -9,9 +9,8 @@
     <ProjectGuid>a421b9ba-2399-4b78-a4a8-02ad3a6cf2f2</ProjectGuid>
     <RootNamespace>DotNetCross.Memory.Unsafe</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">..\artifacts\obj\$(MSBuildProjectName)</BaseIntermediateOutputPath>
-    <OutputPath Condition="'$(OutputPath)'=='' ">..\artifacts\bin\$(MSBuildProjectName)\</OutputPath>
+    <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
   </PropertyGroup>
-
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
   </PropertyGroup>
