+++ src/DotNetCross.Memory.Unsafe/project.json
@@ -1,18 +1,18 @@
 {
     "version": "0.2.3",
 
-    "compilationOptions": {
+    "buildOptions": {
+        "debugType": "portable",
         "allowUnsafe": true
     },
 
-    "dependencies": {
-        "System.Runtime": "4.1.0-rc3-24112-00"
-    },
+    "dependencies": { },
 
-    "frameworks": {
-        "netcore50": {
-            "dependencies": {
-            }
-        }
+  "frameworks": {
+    "netstandard1.0": {
+      "dependencies": {
+        "NETStandard.Library": "1.6.0"
+      }
     }
+  }
 }
