--- src/global.json
@@ -1,6 +1,0 @@
-{
-    "projects":[
-        "DotNetCross.Memory.Unsafe",
-        "DotNetCross.Memory.Unsafe.UnitTests"
-    ]
-}
