+++ tests/DotNetCross.Memory.Unsafe.Benchmark/Properties/AssemblyInfo.cs
