+++ tests/DotNetCross.Memory.Unsafe.Benchmark/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="BenchmarkDotNet" version="0.9.6" targetFramework="net461" />
+  <package id="DotNetCross.Memory.Unsafe" version="0.2.3.2" targetFramework="net461" />
+</packages>
