+++ tests/DotNetCross.Memory.Unsafe.UnitTests/Properties/launchSettings.json
