+++ tests/DotNetCross.Memory.Unsafe.UnitTests/Unsafe.As.Test.cs
