+++ tests/DotNetCross.Memory.Unsafe.UnitTests/Unsafe.AsPointer.Test.cs
