+++ tests/DotNetCross.Memory.Unsafe.UnitTests/Unsafe.InitBlock.Test.cs
