+++ tests/DotNetCross.Memory.Unsafe.UnitTests/Unsafe.Write.Test.cs
