+++ tests/DotNetCross.Memory.Unsafe.UnitTests/Unsafe.WriteRef.Test.cs
