+++ tests/DotNetCross.Memory.Unsafe.UnitTests/project.json
@@ -1,0 +1,35 @@
+{
+    "version": "0.2.3",
+
+    "buildOptions": {
+        "debugType": "portable",
+        "allowUnsafe": true
+    },
+
+    "dependencies": {
+        "dotnet-test-xunit": "2.2.0-preview2-build1029",
+        "DotNetCross.Memory.Unsafe": "0.2.3.4",
+        "xunit": "2.2.0-beta2-build3300"
+    },
+
+    "testRunner": "xunit",
+
+    "frameworks": {
+        "netcoreapp1.0": {
+            "dependencies": {
+                "Microsoft.NETCore.App": {
+                    "type": "platform",
+                    "version": "1.0.0"
+                }
+            },
+            "imports": [
+                "dotnet5.4",
+                "portable-net451+win8"
+            ]
+        }
+    },
+
+    "commands": {
+        "test": "xunit.runner.dnx"
+    }
+}
